Added test cases for ticket service, improved logic for handling the creation of ticket with already existing number, minor changes in the implementation of generic repository

+++ TicketManagementSystem.ApplicationLayer/Interfaces/ITicketService.cs
@@ -6,7 +6,7 @@ namespace TicketManagementSystem.ApplicationLayer.Interfaces
     {
         Task<ServiceResponse<List<TicketDto>>> GetAllAsync();
 
-        Task<ServiceResponse<TicketDto?>> GetByNumberAsync(int number);
+        Task<ServiceResponse<TicketDto?>> GetByNumberAsync(long number);
 
         Task<ServiceResponse<TicketDto?>> CreateAsync(CreateTicketDto createTicketDto);
     }
